+++ include/progress_printer.hpp
@@ -3,15 +3,15 @@
 class ProgressPrinter {
 
   public:
-    ProgressPrinter(const long ne, const double ui);
+    ProgressPrinter(const long long ne, const double ui);
 
-    void operator()(const long index);
+    void operator()(const long long index);
 
   private:
     char *get_time_string() const;
 
     const time_t start_time;
-    const long n_entries;
+    const long long n_entries;
     const double inverse_n_entries;
     const double update_increment;
     double current_percentage;
