+++ source/calibrate_tree.cpp
@@ -30,9 +30,10 @@ int main(int argc, char **argv) {
         tree->Add(input_file.c_str());
     }
 
-    const int first = vm["first"].as<int>();
-    const int last =
-        vm["last"].as<int>() == 0 ? tree->GetEntries() : vm["last"].as<int>();
+    const long long first = vm["first"].as<int>();
+    const long long last = vm["last"].as<long long>() == 0
+                               ? tree->GetEntries()
+                               : vm["last"].as<long long>();
 
     if (first > last) {
         cout << "Error: first entry (" << first
@@ -54,7 +55,7 @@ int main(int argc, char **argv) {
 
     detector_setup.create_branches(tree_calibrated);
 
-    for (int i = first; i <= last; ++i) {
+    for (long long i = first; i <= last; ++i) {
         progress_printer(i - first);
 
         tree->GetEntry(i);
